LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/basic-upgrade.https-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/contentSecurityPolicy/upgrade-insecure-requests/basic-upgrade.https.html requested insecure content from http://127.0.0.1:8443/security/resources/abe.png. This content was automatically upgraded and should be served over HTTPS.


PASS Verify that images are upgraded.
